+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - `View Cargo.toml.orig` button. Allows you to view the original `Cargo.toml` file of a package.
 
+- `--no-default-features` and `--features` command line args to configure feature set to resolve.
+
 ### Changed
 
 - In the main view, the link containing the source folder location was removed in favor of a